BioProfiling.jl API

Table of contents

BioProfiling.jl API

Table of contents

Functions

Index

BioProfiling.jl examples

Introduction

Getting started

Vignette

BioProfiling.jl

Robust multidimensional profiling to quantify complex changes via the minimum covariance determinant.

Introduction

Installation

Learn more

Credits

Functions

Base.filter! — Method

For an experiment e, update in place e.selected_features and e.selected_entries based on an array arr of feature selectors and entry filters. Filters and selectors are applied sequentially.

For an experiment e, update in place e.selected_features and e.selected_entries based on a feature selector s.

For an experiment e, update in place e.selected_features and e.selected_entries based on an entry filter f.

BioProfiling.decorrelate! — Method

Returns column of selected data in Experiment e that are never pairwise-correlated more than 'threshold', prioritizing columns by a given order 'ordercol' (defaults to left to right).

BioProfiling.decorrelate — Method

Returns column of 'data' that are never pairwise-correlated more than 'threshold', prioritizing columns by a given order 'ordercol' (defaults to left to right).

BioProfiling.decorrelate_by_mad! — Method

Returns column of selected data in Experiment e that are never pairwise-correlated more than 'threshold', prioritizing columns with largest median absolute deviation.

BioProfiling.diagnostic — Method

Return information in an experiment e in column(s) feature after filtering with filter f.

BioProfiling.diagnostic_images — Method

Get images in an experiment e whose location is stored in feature s after filtering with filter f. Images will be saved at the path provided if saveimages is set to true, while keeping up to keepsubfolders folders from the original folder structure of the images. If center is true, crosses will indicate the center of selected objects, assuming they are stored in variables :AreaShape_Center_X, :AreaShape_Center_X_1, :AreaShape_Center_Y and :AreaShape_Center_Y_1 (default if you merge nuclear and cytoplasm measurements from CellProfiler outputs). Display up to showlimit images if show is true. Save up to savelimit images if saveimages is true. If rgx provides a list of regex substitutions, it will be applied on all image paths (which is useful if you're in a different file system or environment as the one described in the Experiment's data). If rgb provides a list of 3 lists of regex substitutions, it will be applied to generate the path to 3 images.

BioProfiling.diagnostic_path — Method

Get path to images in an experiment e stored in feature s after filtering with filter f. If center is true, the path will be returned together with the center of the filtered entries, assuming they are stored in variables AreaShape_Center_X, AreaShape_Center_X_1, AreaShape_Center_Y and AreaShape_Center_Y_1. If rgx provides a list of regex substitutions, it will be applied on all file paths in the output (which is useful if you're in a different file system or environment as the one described in the Experiment's data).

BioProfiling.distance_mahalanobis_center — Method

Compute the Mahalanobis Distance to center (MDC) in a dataset 'data' for a given perturbation of indices 'indpert' compared to a reference of indices 'indref'.

BioProfiling.distance_mahalanobis_median — Method

Compute the median Mahalanobis Distance (MD) in a dataset 'data' for a given perturbation of indices 'indpert' compared to a reference of indices 'indref'.

BioProfiling.distance_robust_hellinger — Method

Compute the Robust Hellinger Distance (RHD) in a dataset data for a given perturbation of indices indpert compared to a reference of indices indref.

BioProfiling.distance_robust_mahalanobis_median — Method

Compute the median Robust Mahalanobis Distance (RMD) in a dataset 'data' for a given perturbation of indices 'indpert' compared to a reference of indices 'indref'. See https://e-archivo.uc3m.es/bitstream/handle/10016/24613/ws201710.pdf

BioProfiling.filter_entries! — Method

Filter entries in an Experiment e based on filter(s) f, updating e.selected_entries in place accordingly.

BioProfiling.filter_entries — Method

Return filtered entries in an Experiment e based on filter f

BioProfiling.getdata — Method

Return a copy of the data in Experiment e for its selected entries and features.

BioProfiling.hellinger — Method

Squared hellinger distance for covariance estimators S and centers µ

BioProfiling.logtransform! — Method

Approximate normal distribution of selected entries for all selected features of an Experiment e.

BioProfiling.logtransform — Method

Approximate normal distribution

BioProfiling.mahalanobis — Method

Squared mahalanobis distance for covariance estimator S and center µ

BioProfiling.negation — Method

Return a negative Filter or Selector by inverting the entries or features that are kept and excluded.

BioProfiling.normtransform! — Method

Center and scale all selected entries for each selectead features of an Experiment e on control values matching a Filter f

BioProfiling.normtransform — Method

Center and scale on control values

BioProfiling.robust_morphological_perturbation_value — Function

Compute the Robust Morphological Perturbation Value (RMPV) for a given Experiment e, for all levels of a column s, compared to rows matching a given filter f or where s is equal to ref. The RMPV quantifies the significance of changes between all conditions (levels in s) and a reference condition (defined by the filter f). In brief, the distance of type dist between points of each perturbation and points of the reference is computed and its statistical significance is defined using a permutation test in which the perturbation and reference labels are shuffled nb_rep times. If process_pool is a pool of worker processes, they will be used for parallel computation in the permutation test. This returns a DataFrame with three columns:

Condition: the levels in s

Distance: the distance between a condition and the

reference

RMPV: the RMPV (empirical p-value corrected for multiple

testing)

BioProfiling.select_features! — Method

Return selected features in an Experiment e based on selectors s, updating e.selected_features in place accordingly.

BioProfiling.select_features — Method

Return selected features in an Experiment e based on selectors s

BioProfiling.shuffled_distance_mahalanobis_center — Method

Permute labels and compute the Mahalanobis Distance to center (MDC) in a dataset 'data' for a given perturbation of indices 'indpert' compared to a reference of indices 'indref', to create an empirical distribution.

BioProfiling.shuffled_distance_mahalanobis_median — Method

Permute labels and compute the median Mahalanobis Distance (RMD) in a dataset 'data' for a given perturbation of indices 'indpert' compared to a reference of indices 'indref', to create an empirical distribution.

BioProfiling.shuffled_distance_robust_hellinger — Method

Permute labels and compute the Robust Hellinger Distance (RHD) in a dataset data for a given perturbation of indices indpert compared to a reference of indices indref, to create an empirical distribution.

BioProfiling.shuffled_distance_robust_mahalanobis_median — Method

Permute labels and compute the median Robust Mahalanobis Distance (RMD) in a dataset 'data' for a given perturbation of indices 'indpert' compared to a reference of indices 'indref', to create an empirical distribution.

DataFrames.select! — Method

For an experiment e, update in place e.selected_features and e.selected_entries based on an array arr of feature selectors and entry filters. Filters and selectors are applied sequentially.

For an experiment e, update in place e.selected_features and e.selected_entries based on a feature selector s.

For an experiment e, update in place e.selected_features and e.selected_entries based on an entry filter f.

Index

Base.filter!

BioProfiling.decorrelate

BioProfiling.decorrelate!

BioProfiling.decorrelate_by_mad!

BioProfiling.diagnostic

BioProfiling.diagnostic_images

BioProfiling.diagnostic_path

BioProfiling.distance_mahalanobis_center

BioProfiling.distance_mahalanobis_median

BioProfiling.distance_robust_hellinger

BioProfiling.distance_robust_mahalanobis_median

BioProfiling.filter_entries

BioProfiling.filter_entries!

BioProfiling.getdata

BioProfiling.hellinger

BioProfiling.logtransform

BioProfiling.logtransform!

BioProfiling.mahalanobis

BioProfiling.negation

BioProfiling.normtransform

BioProfiling.normtransform!

BioProfiling.robust_morphological_perturbation_value

BioProfiling.select_features

BioProfiling.select_features!

BioProfiling.shuffled_distance_mahalanobis_center

BioProfiling.shuffled_distance_mahalanobis_median

BioProfiling.shuffled_distance_robust_hellinger

BioProfiling.shuffled_distance_robust_mahalanobis_median

DataFrames.select!